requirements.txt
Backtesting==0.3.1
bidict==0.21.2
bokeh==2.2.3
certifi==2020.12.5
chardet==4.0.0
configparser==5.0.1
cycler==0.10.0
fxcmpy==1.2.6
idna==2.10
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
MetaTrader5==5.0.33
numpy==1.19.5
packaging==20.9
pandas==1.2.1
Pillow==8.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pyti==0.2.2
PyYAML==5.4.1
tornado==6.1
typing-extensions==3.7.4.3
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
configparser==5.0.1
fxcmpy==1.2.6
idna==2.10
pandas==1.2.1
PySocks==1.7.1
python-engineio==3.14.2
python-socketio==4.6.1
pytz==2021.1
requests==2.25.1
six==1.15.0
socketIO-client==0.7.2
urllib3==1.26.3
websocket-client==0.57.0

